Change integers in RobotMap to doubles for decimal notations.

+++ src/org/usfirst/frc/team694/robot/RobotMap.java
@@ -69,8 +69,8 @@ public final class RobotMap {
     /******************************************************************************
      * Line Sensor Constants
      *****************************************************************************/
-    public static final int DRIVETRAIN_LINE_SENSOR_INITIALIZE_TIME = 3;
-    public static final int DRIVETRAIN_LINE_SENSOR_THRESHOLD = -1;
+    public static final double DRIVETRAIN_LINE_SENSOR_INITIALIZE_TIME = 3;
+    public static final double DRIVETRAIN_LINE_SENSOR_THRESHOLD = -1.0;
     
     /******************************************************************************
      * Drivetrain Encoder/Movement Constants
